Added translations and debug change
Added ability for translations with new static Language class in Proj.Modules.Misc which loads a .lang file from "src/resources/*asset_pack/data/lang/*language.lang" in a custom format atm because i couldnt figure out how to do json. This is then saved in a Dictionary<String, String> where you can then use the Text method in Language to get the value and returns the id if its invalid for a failsafe.

I also changed the debug to not require Debug.get() every time as it got very annoying and i was dumb when i implemented it because i forgot Debug.send() is sent from the same place so i just got the method in there instead.